requirements.txt
streamlit>=1.0.0
ollama>=0.1.0
pillow>=9.0.0
pandas>=1.0.0 